LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-pointer-remove-text-expected.txt


FAIL pointer updates (removing text nodes) assert_equals: error events on c expected 1 but got 0
